@@ -1,14 +1,13 @@
 plugins {
-	id 'fabric-loom' version '0.11-SNAPSHOT'
+	id 'fabric-loom' version '1.0-SNAPSHOT'
 	id 'maven-publish'
 	id "io.freefair.lombok" version "6.0.0-m2"
-	id "com.diffplug.spotless" version "5.17.1"
 }
 
 apply from: 'https://raw.githubusercontent.com/DarkKronicle/GradleScripts/main/gradle/publish.gradle'
 
-sourceCompatibility = JavaVersion.VERSION_16
-targetCompatibility = JavaVersion.VERSION_16
+sourceCompatibility = JavaVersion.VERSION_17
+targetCompatibility = JavaVersion.VERSION_17
 
 archivesBaseName = project.archives_base_name
 version = "${project.minecraft_version}-${project.mod_version}"
@@ -22,6 +21,7 @@ if (file("extra.properties").exists()) {
 repositories {
 	mavenCentral()
 	maven { url 'https://masa.dy.fi/maven' }
+    maven { url 'https://maven.terraformersmc.com/releases/' }
 	maven { url 'https://jitpack.io' }
 }
 
@@ -30,8 +30,8 @@ dependencies {
 	mappings "net.fabricmc:yarn:${project.yarn_mappings}:v2"
 	modImplementation "net.fabricmc:fabric-loader:${project.loader_version}"
 	modImplementation "net.fabricmc.fabric-api:fabric-api:${project.fabric_api_version}"
-	modImplementation "fi.dy.masa.malilib:malilib-fabric-1.19.0:${project.malilib_version}"
-	modCompileOnly "io.github.prospector:modmenu:1.16.8"
+	modImplementation "fi.dy.masa.malilib:malilib-fabric-1.19.3:${project.malilib_version}"
+	modCompileOnly "com.terraformersmc:modmenu:5.0.1"
 
 	implementation "com.github.DarkKronicle.Konstruct:addons:${project.konstruct_version}"
 	implementation "com.github.DarkKronicle.Konstruct:core:${project.konstruct_version}"
@@ -87,19 +87,6 @@ tasks.register('downloadExtra') {
 
 assemble {
 	dependsOn 'downloadExtra'
-}
-
-spotless {
-	ratchetFrom 'origin/main'
-	format 'misc', {
-		// define the files to apply `misc` to
-		target '*.gradle', '*.md', '.gitignore'
-
-		// define the steps to apply to those files
-		trimTrailingWhitespace()
-		indentWithTabs() // or spaces. Takes an integer argument if you don't like 4
-		endWithNewline()
-	}
 }
 
 // configure the maven publication
