@@ -9,12 +9,12 @@ package io.github.darkkronicle.advancedchatcore.chat;
 
 import io.github.darkkronicle.advancedchatcore.AdvancedChatCore;
 import io.github.darkkronicle.advancedchatcore.interfaces.IStringFilter;
+import net.minecraft.client.MinecraftClient;
+import org.apache.logging.log4j.Level;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.text.Text;
-import org.apache.logging.log4j.Level;
 
 public class MessageSender {
 
@@ -38,10 +38,6 @@ public class MessageSender {
     }
 
     public void sendMessage(String string) {
-        sendMessage(string, null);
-    }
-
-    public void sendMessage(String string, Text text) {
         String unfiltered = string;
         for (IStringFilter filter : filters) {
             Optional<String> filtered = filter.filter(string);
@@ -61,9 +57,9 @@ public class MessageSender {
 
         if (client.player != null) {
             if (string.startsWith("/")) {
-                this.client.player.sendCommand(string.substring(1), text);
+                this.client.getNetworkHandler().sendChatCommand(string.substring(1));
             } else {
-                this.client.player.sendChatMessage(string, text);
+                this.client.getNetworkHandler().sendChatMessage(string);
             }
         }
     }
